+++ src/main/java/com/gmail/kramarenko104/controller/OrderController.java
@@ -17,11 +17,13 @@ import org.springframework.web.bind.annotation.RequestParam;
 @RequestMapping("/order")
 public class OrderController {
 
+    private static Logger logger = Logger.getLogger(OrderController.class);
+    private static final String DB_WARNING = "Check your connection to DB!";
+
     @Autowired
     private OrderService orderService;
     @Autowired
     private CartService cartService;
-    private static Logger logger = Logger.getLogger(OrderController.class);
 
     public OrderController() {
     }
@@ -31,34 +33,39 @@ public class OrderController {
         return "order";
     }
 
-    @RequestMapping(method = RequestMethod.POST, produces="text/json")
+    @RequestMapping(method = RequestMethod.POST, produces = "text/json")
     public String doPost(@RequestParam("action") String action,
-                            @RequestParam("userId") int userId,
-                            Model model) {
+                         @RequestParam("userId") int userId,
+                         Model model) {
 
         String jsondata = null;
-        if (model.asMap().get("user") != null) {
-            // get info from Ajax POST request (updateCart.js)
-            if (action != null && (action.equals("makeOrder"))) {
-                logger.debug("OrderServlet.POST: got userId from POST request: " + userId);
 
-                // any user can have only one existing now cart and many processed orders (userId uniquely identifies cart)
-                Cart cart = cartService.getCart(userId);
-                logger.debug("OrderServlet.POST: got cart from DB: " + cart);
+        if (cartService.sessionIsOpen()) {
+            if (model.asMap().get("user") != null) {
+                // get info from Ajax POST request (updateCart.js)
+                if (action != null && (action.equals("makeOrder"))) {
+                    logger.debug("OrderServlet.POST: got userId from POST request: " + userId);
 
-                // order will be created based on the cart's content
-                Order newOrder = orderService.createOrder(userId, cart.getProducts());
-                logger.debug("OrderServlet.POST: !!! new Order was created: " + newOrder);
-                model.addAttribute("newOrder", newOrder);
+                    // any user can have only one existing now cart and many processed orders (userId uniquely identifies cart)
+                    Cart cart = cartService.getCart(userId);
+                    logger.debug("OrderServlet.POST: got cart from DB: " + cart);
 
-                // send JSON back with the new Order to show on order.jsp
-                if (newOrder != null) {
-                    jsondata = new Gson().toJson(newOrder);
-                    logger.debug("OrderServlet: send JSON data to cart.jsp ---->" + jsondata);
+                    // order will be created based on the cart's content
+                    Order newOrder = orderService.createOrder(userId, cart.getProducts());
+                    logger.debug("OrderServlet.POST: !!! new Order was created: " + newOrder);
+                    model.addAttribute("newOrder", newOrder);
+
+                    // send JSON back with the new Order to show on order.jsp
+                    if (newOrder != null) {
+                        jsondata = new Gson().toJson(newOrder);
+                        logger.debug("OrderServlet: send JSON data to cart.jsp ---->" + jsondata);
+                    }
+                    cartService.deleteCart(Integer.valueOf(userId));
+                    model.addAttribute("userCart", null);
                 }
-                cartService.deleteCart(Integer.valueOf(userId));
-                model.addAttribute("userCart", null);
             }
+        } else { // connection to DB is closed
+            model.addAttribute("warning", DB_WARNING);
         }
         return jsondata;
     }
